+++ e2e/cli-e2e/mocks/fixtures/code-pushup.config.coverage.ts
@@ -15,35 +15,17 @@ export default {
       title: 'Code coverage',
       refs: [
         {
-          type: 'audit',
+          type: 'group',
           plugin: 'coverage',
-          slug: 'function-coverage',
-          weight: 1,
-        },
-        {
-          type: 'audit',
-          plugin: 'coverage',
-          slug: 'branch-coverage',
-          weight: 1,
-        },
-        {
-          type: 'audit',
-          plugin: 'coverage',
-          slug: 'line-coverage',
+          slug: 'coverage',
           weight: 1,
         },
       ],
     },
   ],
   plugins: [
-    coveragePlugin({
-      coverageTypes: ['branch', 'function', 'line'],
-      reports: [
-        {
-          resultsPath: join('e2e', 'cli-e2e', 'mocks', 'fixtures', 'lcov.info'),
-          pathToProject: join('packages', 'cli'),
-        },
-      ],
+    await coveragePlugin({
+      reports: [join('e2e', 'cli-e2e', 'mocks', 'fixtures', 'lcov.info')],
     }),
   ],
 } satisfies CoreConfig;