+++ examples/plugins/vite.config.unit.ts
@@ -10,11 +10,15 @@ export default defineConfig({
     cache: {
       dir: '../../node_modules/.vitest',
     },
+    coverage: {
+      reporter: ['lcov'],
+    },
     environment: 'node',
     include: ['src/**/*.unit.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     globalSetup: ['global-setup.ts'],
     setupFiles: [
       '../../testing-utils/src/lib/setup/fs.mock.ts',
+      '../../testing-utils/src/lib/setup/git.mock.ts',
       '../../testing-utils/src/lib/setup/console.mock.ts',
       '../../testing-utils/src/lib/setup/reset.mocks.ts',
     ],