+++ packages/cli/vite.config.unit.ts
@@ -18,6 +18,7 @@ export default defineConfig({
     globalSetup: ['global-setup.ts'],
     setupFiles: [
       '../../testing-utils/src/lib/setup/fs.mock.ts',
+      '../../testing-utils/src/lib/setup/git.mock.ts',
       '../../testing-utils/src/lib/setup/console.mock.ts',
       '../../testing-utils/src/lib/setup/portal-client.mock.ts',
       '../../testing-utils/src/lib/setup/reset.mocks.ts',