+++ packages/plugin-coverage/src/lib/runner/lcov/lcov-runner.integration.test.ts
@@ -1,10 +1,11 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { describe, it } from 'vitest';
-import { lcovResultsToAuditOutputs } from './runner';
+import { osAgnosticAuditOutputs } from '@code-pushup/testing-utils';
+import { lcovResultsToAuditOutputs } from './lcov-runner';
 
 describe('lcovResultsToAuditOutputs', () => {
-  it('should correctly convert lcov results to AuditOutputs', async () => {
+  it('should correctly convert lcov results to AuditOutputs and prepend project paths', async () => {
     /**
      * The stats passed in the fixture are as follows
      * Functions: 2 found, 2 covered (100% coverage)
@@ -23,11 +24,11 @@ describe('lcovResultsToAuditOutputs', () => {
             'mocks',
             'single-record-lcov.info',
           ),
-          pathToProject: join('packages', 'cli'),
+          pathToProject: 'packages/cli',
         },
       ],
       ['branch', 'function', 'line'],
     );
-    expect(results).toMatchSnapshot();
+    expect(osAgnosticAuditOutputs(results)).toMatchSnapshot();
   });
 });