+++ testing-utils/src/lib/utils/os-agnostic-paths.ts
@@ -1,0 +1,40 @@
+import type {
+  AuditOutput,
+  AuditOutputs,
+  AuditReport,
+} from '@code-pushup/models';
+
+export function osAgnosticPath(path: string): string {
+  return path.replace(process.cwd(), '<CWD>').replace(/\\/g, '/');
+}
+
+export function osAgnosticAudit<T extends AuditOutput | AuditReport>(
+  audit: T,
+): T {
+  if (
+    !audit.details?.issues.length ||
+    audit.details.issues.every(issue => issue.source == null)
+  ) {
+    return audit;
+  }
+  return {
+    ...audit,
+    details: {
+      issues: audit.details.issues.map(issue =>
+        issue.source == null
+          ? issue
+          : {
+              ...issue,
+              source: {
+                ...issue.source,
+                file: osAgnosticPath(issue.source.file),
+              },
+            },
+      ),
+    },
+  };
+}
+
+export function osAgnosticAuditOutputs(audits: AuditOutputs): AuditOutputs {
+  return audits.map(osAgnosticAudit);
+}